+++ sshd-sftp/src/test/java/org/apache/sshd/sftp/client/fs/SftpFileSystemTest.java
@@ -1081,7 +1081,7 @@ public class SftpFileSystemTest extends AbstractSftpFilesSystemSupport {
                 }
 
                 assertFalse(found, "Multiple configurations: " + line);
-                assertEquals(SftpFileSystemProvider.class.getName(), line, "Mismatched configuration");
+                assertEquals(SftpFileSystemProviderFacade.class.getName(), line, "Mismatched configuration");
                 found = true;
             }
         }